namespaces Loan and Borrowe renamed
Domain.Loan and Domain.Borrower are renamed into respectively Domain.Loans and Domain.Borrowers
