+++ ApiTP1MethodeDeDev/ApiTP1MethodeDeDev/Views/Borrowers/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model Domain.Borrower.Borrower
+﻿@model Domain.Borrowers.Borrower
 
 @{
     ViewData["Title"] = "Details";
