+++ ApiTP1MethodeDeDev/ApiTP1MethodeDeDev/Views/Borrowers/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<Domain.Borrower.Borrower>
+﻿@model IEnumerable<Domain.Borrowers.Borrower>
 
 @{
     ViewData["Title"] = "Index";
