+++ ApiTP1MethodeDeDev/Domain/Borrowers/Borrower.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Borrower
+namespace Domain.Borrowers
 {
     public class Borrower
     {
