+++ ApiTP1MethodeDeDev/Domain/Borrowers/IBorrowerRepository.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Borrower
+namespace Domain.Borrowers
 {
     public interface IBorrowerRepository
     {
