+++ ApiTP1MethodeDeDev/Infrastructure/BorrowerEntity.cs
@@ -5,7 +5,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Domain.Borrower;
+using Domain.Borrowers;
 
 namespace Infrastructure
 {
