+++ ApiTP1MethodeDeDev/Infrastructure/InMemoryLoanRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Domain.Loan;
+using Domain.Loans;
 
 namespace Infrastructure
 {
